+++ test/global_test.go
@@ -5,6 +5,7 @@ package helmtest
 
 import (
 	"bytes"
+	"errors"
 	"fmt"
 	"io"
 	"reflect"
@@ -134,4 +135,23 @@ func ensureVolumePresent(matchVol v1.Volume, volumes []v1.Volume) bool {
 		}
 	}
 	return present
+}
+
+func GetEnvVarByName(envVars []v1.EnvVar, name string) (error, string) {
+	for _, v := range envVars {
+		if v.Name == name {
+			return nil, v.Value
+		}
+	}
+	return errors.New("Not found"), ""
+}
+
+// GetContainerByName returns container or nil
+func GetContainerByName(name string, containers []v1.Container) *v1.Container {
+	for _, c := range containers {
+		if c.Name == name {
+			return &c
+		}
+	}
+	return nil
 }
